기존 코드 리팩토링
적 생성 성장 통합 관리, 리그 매니저 역할 축소
