+++ Main_Project/Assets/Scripts/Data/Enemy/EnemyTeamService.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 48b7e8358a151464b86bd7fa9a5539b7
+guid: 599164a5e2ea44ea3bef148084e121c8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
